+++ pages/login/index.tsx
@@ -1,4 +1,4 @@
-import Login from "@/components/Login";
+import Login from "@/container/Login";
 
 export default function LoginPage() {
   return (
